pyproject.toml
[project]
name = "learning-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chroma>=0.2.0",
    "chromadb>=0.5.20",
    "langchain>=0.3.15",
    "langchain-chroma>=0.2.0",
    "langchain-community>=0.3.15",
    "langchain-openai>=0.3.1",
    "langchain-text-splitters>=0.3.5",
    "jupyterlab>=1.1.1",
    "psycopg2-binary>=2.9.10",
    "langchain-postgres>=0.0.12",
    "langchain-huggingface>=0.1.2",
    "tf-keras>=2.18.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest-cov>=5.0.0",
    "pytest-datadir>=1.5.0",
    "pytest-flake8>=1.1.1",
    "pytest-instafail>=0.5.0",
    "pytest-json>=0.4.0",
    "pytest-leaks>=0.3.1",
    "pytest-pycodestyle>=2.3.1",
    "pytest-redis>=3.0.2",
    "pytest-regressions>=2.5.0",
    "pytest-repeat>=0.9.3",
    "pytest-replay>=1.5.1",
    "pytest-rerunfailures>=14.0",
    "pytest-selenium>=4.1.0",
    "pytest-sugar>=1.0.0",
    "pytest-timeout>=2.3.1",
    "pytest-vagrant>=2.1.0",
    "pytest-vcr>=1.0.2",
    "pytest-xdist>=3.5.0",
    "pytest-xvfb>=3.0.0",
]

